Break up test_utils.rs: extract `DictStateReader`
and `CachedState` constructors

commit-id:dc18fb72
