+++ crates/blockifier/src/test_utils/dict_state_reader.rs
@@ -1,0 +1,63 @@
+use std::collections::HashMap;
+
+use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
+use starknet_api::hash::StarkFelt;
+use starknet_api::state::StorageKey;
+
+use crate::execution::contract_class::ContractClass;
+use crate::state::cached_state::ContractStorageKey;
+use crate::state::errors::StateError;
+use crate::state::state_api::{StateReader, StateResult};
+
+/// A simple implementation of `StateReader` using `HashMap`s as storage.
+#[derive(Debug, Default)]
+pub struct DictStateReader {
+    pub storage_view: HashMap<ContractStorageKey, StarkFelt>,
+    pub address_to_nonce: HashMap<ContractAddress, Nonce>,
+    pub address_to_class_hash: HashMap<ContractAddress, ClassHash>,
+    pub class_hash_to_class: HashMap<ClassHash, ContractClass>,
+    pub class_hash_to_compiled_class_hash: HashMap<ClassHash, CompiledClassHash>,
+}
+
+impl StateReader for DictStateReader {
+    fn get_storage_at(
+        &mut self,
+        contract_address: ContractAddress,
+        key: StorageKey,
+    ) -> StateResult<StarkFelt> {
+        let contract_storage_key = (contract_address, key);
+        let value = self.storage_view.get(&contract_storage_key).copied().unwrap_or_default();
+        Ok(value)
+    }
+
+    fn get_nonce_at(&mut self, contract_address: ContractAddress) -> StateResult<Nonce> {
+        let nonce = self.address_to_nonce.get(&contract_address).copied().unwrap_or_default();
+        Ok(nonce)
+    }
+
+    fn get_compiled_contract_class(
+        &mut self,
+        class_hash: &ClassHash,
+    ) -> StateResult<ContractClass> {
+        let contract_class = self.class_hash_to_class.get(class_hash).cloned();
+        match contract_class {
+            Some(contract_class) => Ok(contract_class),
+            _ => Err(StateError::UndeclaredClassHash(*class_hash)),
+        }
+    }
+
+    fn get_class_hash_at(&mut self, contract_address: ContractAddress) -> StateResult<ClassHash> {
+        let class_hash =
+            self.address_to_class_hash.get(&contract_address).copied().unwrap_or_default();
+        Ok(class_hash)
+    }
+
+    fn get_compiled_class_hash(
+        &mut self,
+        class_hash: ClassHash,
+    ) -> StateResult<starknet_api::core::CompiledClassHash> {
+        let compiled_class_hash =
+            self.class_hash_to_compiled_class_hash.get(&class_hash).copied().unwrap_or_default();
+        Ok(compiled_class_hash)
+    }
+}
